refactor: remove unused fund-escrow hook and update import path in fund-escrow-dialog hook
